backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
pillow==11.2.1
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
